lk2nd: smp: Add SMP spin table for booting CPU cores
On ARM, the Power State Coordination Interface (PSCI) is the standard
way to boot secondary CPU cores, by making calls into the TrustZone
firmware. Unfortunately, it was adopted quite late by Qualcomm. Many
older firmware versions and platforms lack PSCI support.

While Linux ARM32 still supports custom CPU "enable-method"s, this is
not supported anymore on Linux ARM64. Only PSCI or "spin-table" is
allowed there. "spin-table" works by booting the CPU cores as part of
the boot firmware, and letting them "spin" while checking if a release
address was written by the operating system.

The advantage is that this can be implemented easily as part of lk2nd,
while implementing PSCI is not as simple (or perhaps impossible) since
the original TrustZone firmware is closed-source and usually signed.

The new "smp" module added in this commit automatically detects if PSCI
is supported (or if the device tree already uses a custom "enable-method"
on ARM32). Otherwise the CPUs are booted and the DT is modified to use
"spin-table" instead.

MSM8916/MSM8939 are the main target of this code, but older ARM32-only
platforms are supported as well to simplify booting the CPU cores in
operating systems other than Linux.

This is a squashed and slightly reworked version of similar code in the
old lk2nd "master" branch. Full commit history is available there.